# docs/openapi.yaml
openapi: 3.1.0
info:
  title: muse-replicate Managed Agents API
  version: 0.1.0
  description: |
    企业私有化 Managed Agents 平台（Meta Muse 复刻）的 HTTP API，方言为 BigModel Managed Agents 子集。

    **鉴权**：`Authorization: Bearer mas_sk_...` 或 `x-api-key: mas_sk_...`（任选其一）。

    **方言 headers**（建议随请求发送）：
    - `zai-version: 2026-05-26`
    - `zai-beta: managed-agents-2026-05-26`

    兼容 Anthropic 方言（`anthropic-version` / `anthropic-beta` / `x-api-key`），服务端按 header 自动检测方言并适配错误信封。

    **错误信封**（所有非 2xx）：
    ```json
    { "type": "error", "error": { "type": "invalid_request_error", "message": "...", "details": {} }, "request_id": "req_..." }
    ```

    **通用语义**：
    - 所有 POST 接受 `Idempotency-Key` header（同 key 同 body 回放首次响应，异 body 409 `idempotency_conflict`）；
    - 列表接口支持 `limit` / 分页游标（`next_page`），Files 用 `before_id`/`after_id`；
    - 事件日志是唯一事实来源：会话事件按 `seq` 单调定序；SSE 见 `GET /v1/sessions/{id}/events/stream`。
servers:
  - url: http://127.0.0.1:8080
    description: 本地部署（PORT 环境变量覆盖）
...
security:
  - bearerAuth: []
  - apiKeyHeader: []
paths:
...
  /v1/skills:
    post:
      tags: [Skills]
      operationId: uploadSkill
      summary: multipart 上传 Skill zip（SKL-01/02/03；根 SKILL.md；≤200 文件静默截断）
      parameters: *dialect
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                directory: { type: string, description: "缺省取 zip 公共根目录" }
                description: { type: string }
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Skill" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      tags: [Skills]
      operationId: listSkills
      summary: Skill 列表（source=user|zai；SKL-07）
      parameters:
        - $ref: "#/components/parameters/ZaiVersion"
        - $ref: "#/components/parameters/ZaiBeta"
        - { name: source, in: query, schema: { type: string, enum: [user, zai] } }
        - { name: include_archived, in: query, schema: { type: string, enum: ["true"] } }
      responses:
        "200":
          description: Skill 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Skill" } }
                  next_page: { type: [string, "null"] }
  /v1/skills/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [Skills]
      operationId: getSkill
      summary: 查询 Skill
      parameters: *dialectGet
      responses:
        "200":
          description: Skill
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Skill" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Skills]
      operationId: deleteSkill
      summary: 删除 Skill（仍被 agent 版本引用/会话挂载 409；SKL-05）
      parameters: *dialectGet
      responses:
        "200": { description: 删除成功 }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/skills/{id}/archive:
    post:
      tags: [Skills]
      operationId: archiveSkill
      summary: 归档 Skill（幂等）
      parameters: *dialect
      responses:
        "200":
          description: 归档后的 Skill
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Skill" }
  /v1/skills/{id}/versions:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [Skills]
      operationId: listSkillVersions
      summary: 版本列表（desc）
      parameters: *dialectGet
      responses:
        "200":
          description: 版本列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/SkillVersion" } }
                  next_page: { type: [string, "null"] }
    post:
      tags: [Skills]
      operationId: uploadSkillVersion
      summary: 追加新版本（SKL-04）
      parameters: *dialect
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: 新版本
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SkillVersion" }
  /v1/skills/{id}/content:
    get:
      tags: [Skills]
      operationId: downloadSkillContent
      summary: 下载最新版本 zip
      parameters: *dialectGet
      responses:
        "200":
          description: application/zip
          content:
            application/zip:
              schema: { type: string, format: binary }
  /v1/skills/{id}/versions/{version}/content:
    get:
      tags: [Skills]
      operationId: downloadSkillVersionContent
      summary: 下载指定版本 zip（SKL-03/04）
      parameters:
        - $ref: "#/components/parameters/ZaiVersion"
        - $ref: "#/components/parameters/ZaiBeta"
        - { name: version, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        "200":
          description: application/zip
          content:
            application/zip:
              schema: { type: string, format: binary }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Authorization: Bearer mas_sk_..."
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
      description: "x-api-key: mas_sk_...（BigModel/Anthropic 方言兼容）"
  parameters:
    ZaiVersion:
      name: zai-version
      in: header
      required: false
      schema: { type: string, default: "2026-05-26" }
      description: BigModel API 版本（等价 anthropic-version）
    ZaiBeta:
      name: zai-beta
      in: header
      required: false
      schema: { type: string, default: "managed-agents-2026-05-26" }
      description: Managed Agents beta 通道（等价 anthropic-beta）
...
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: 请求体/参数不合法（invalid_request_error）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    NotFound:
      description: 资源不存在（not_found_error）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: 状态冲突 / 乐观锁 / 幂等冲突 / 引用约束（conflict_error）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    Error:
      type: object
      required: [type, message]
      properties:
        type:
          type: string
          enum: [invalid_request_error, authentication_error, permission_error, not_found_error, conflict_error, rate_limit_error, api_error, overloaded_error]
        message: { type: string }
        details:
          type: [object, "null"]
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [type, error, request_id]
      properties:
        type: { const: error }
        error: { $ref: "#/components/schemas/Error" }
        request_id: { type: string }
    Timestamp:
      type: string
      format: date-time
...
    Skill:
      type: object
      required: [id, type, directory]
      properties:
        id: { type: string }
        type: { const: skill }
        directory: { type: string }
        description: { type: [string, "null"] }
        source: { type: string, enum: [user, zai] }
        latest_version: { type: integer, minimum: 1 }
        created_at: { $ref: "#/components/schemas/Timestamp" }
        updated_at: { $ref: "#/components/schemas/Timestamp" }
        archived_at: { type: [string, "null"], format: date-time }
    SkillVersion:
      type: object
      required: [id, type, skill_id, version]
      properties:
        id: { type: string }
        type: { const: skill_version }
        skill_id: { type: string }
        version: { type: integer, minimum: 1 }
        sha256: { type: string }
        bytes: { type: integer }
        created_at: { $ref: "#/components/schemas/Timestamp" }
...